feat: Added setAdditionalRecipeInformation method
